pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cloud-finops-env"
version = "1.0.0"
description = "OpenEnv environment for cloud infrastructure cost optimization"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Cloud FinOps Optimizer" }
]
keywords = ["openenv", "reinforcement-learning", "finops", "cloud", "rl"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "openai>=1.51.0",
    "requests>=2.32.0",
    "python-multipart>=0.0.12",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cloud_finops_env*"]

[project.scripts]
finops-server = "server.app:app"
